langgraph-vs-crewai/pyproject.toml
[project]
name = "langgraph-vs-crewai"
version = "1.0.0"
description = "Framework Comparison Benchmark Suite"
authors = [
    { name = "Ismail Sajid" }
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "matplotlib>=3.7.0",
    "jinja2>=3.1.0",
    "scipy>=1.10.0",
    "docker>=6.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "structlog>=23.1.0",
]

[project.scripts]
benchmark = "benchmark.cli:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
